Keep shared streaming helper in existing module
